Add base class for metrics (#65)
