+++ test/metrics-rabbitmq-exchange_spec.rb
@@ -47,14 +47,14 @@ describe RabbitMQExchangeMetrics, 'run' do
   end
 
   it "should output nothing and return ok when there are no exchanges" do
-    allow(check).to receive(:acquire_rabbitmq_exchanges).and_return []
+    allow(check).to receive(:acquire_rabbitmq_info).and_return []
     expect(check).not_to receive(:output)
     expect(check).to receive(:ok)
     check.run
   end
 
   it "should by default output all exchanges and all metrics" do
-    allow(check).to receive(:acquire_rabbitmq_exchanges).and_return [e1, e2]
+    allow(check).to receive(:acquire_rabbitmq_info).and_return [e1, e2]
 
     expect(check).to receive(:output).with(/.+.rabbitmq.e1.name$/, 'e1', timestamp)
     expect(check).to receive(:output).with(/.+.rabbitmq.e1.type$/, 'topic', timestamp)
@@ -72,7 +72,7 @@ describe RabbitMQExchangeMetrics, 'run' do
 
   it "should output only the exchanges specified by the filter option" do
     check.config[:filter] = '.2'
-    allow(check).to receive(:acquire_rabbitmq_exchanges).and_return [e1, e2]
+    allow(check).to receive(:acquire_rabbitmq_info).and_return [e1, e2]
     expect(check).not_to receive(:output).with(/e1/, any_args)
     expect(check).to receive(:output).with(/e2/, any_args).exactly(e2.size).times
     expect(check).to receive(:ok)
@@ -81,7 +81,7 @@ describe RabbitMQExchangeMetrics, 'run' do
 
   it "should output only the metrics specified by the metrics option" do
     check.config[:metrics] = 'message_stats'
-    allow(check).to receive(:acquire_rabbitmq_exchanges).and_return [e1]
+    allow(check).to receive(:acquire_rabbitmq_info).and_return [e1]
     expect(check).to receive(:output).with(/e1.message_stats.publish$/, 100, timestamp)
     expect(check).to receive(:output).with(/e1.message_stats.ack$/, 50, timestamp)
     expect(check).to receive(:ok)
